requirements.txt
# Python dependencies
discord.py
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
python-dotenv
